Update solution to MvvmCross 6.2.2 Update project to use only one .netStandard
