+++ MvvX.Plugins.Caching.Droid.UnitTests/Properties/AndroidManifest.xml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android" package="MvvX.Plugins.Caching.Droid.UnitTests.MvvX.Plugins.Caching.Droid.UnitTests" android:versionCode="1" android:versionName="1.0">
-	<uses-sdk android:minSdkVersion="16" />
+	<uses-sdk android:minSdkVersion="16" android:targetSdkVersion="28" />
 	<application android:label="MvvX.Plugins.Caching.Droid.UnitTests"></application>
 </manifest>
