+++ MvvX.Plugins.Caching.Droid.UnitTests/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MvvmCross" version="6.2.3" targetFramework="monoandroid90" />
+</packages>
