--- MvvX.Plugins.Caching.Droid/Plugin.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using MvvmCross.Platform;
-using MvvmCross.Platform.Plugins;
-
-namespace MvvX.Plugins.Caching.Droid
-{
-    public class Plugin : IMvxConfigurablePlugin
-    {
-        private string configurationKey;
-
-        public void Configure(IMvxPluginConfiguration configuration)
-        {
-            var config = configuration as PluginConfiguration;
-            if (config == null || string.IsNullOrWhiteSpace(config.GlobalCacheKey))
-                throw new ArgumentNullException("configuration is invalid. null or globalKey is null or white spaces");
-
-            configurationKey = config.GlobalCacheKey;
-        }
-
-        public void Load()
-        {
-            Mvx.RegisterSingleton<ICacheManager>(new CacheManager(configurationKey));
-        }
-    }
-}
