--- MvvX.Plugins.Caching.Droid/System.Runtime.Caching/CacheEntryChangeMonitor.cs
@@ -1,10 +1,0 @@
-
-using System.Collections.Generic;
-
-namespace MvvX.Plugins.Caching.Droid.System.Runtime.Caching
-{
-    internal class CacheEntryChangeMonitor
-    {
-        public IList<string> DependencyKeys { get; set; }
-    }
-}
