--- MvvX.Plugins.Caching.Droid/System.Runtime.Caching/CacheItemPolicy.cs
@@ -1,18 +1,0 @@
-
-using System;
-using System.Collections.Generic;
-
-namespace MvvX.Plugins.Caching.Droid.System.Runtime.Caching
-{
-    internal class CacheItemPolicy
-    {
-        public DateTimeOffset AbsoluteExpiration { get; set; }
-
-        public IList<CacheEntryChangeMonitor> ChangeMonitors { get; internal set; }
-
-        public CacheItemPolicy()
-        {
-            ChangeMonitors = new List<CacheEntryChangeMonitor>();
-        }
-    }
-}
