+++ MvvX.Plugins.Caching.Wpf.UnitTests/CacheManagerTest.cs
@@ -15,18 +15,16 @@ namespace MvvX.Plugins.Caching.Wpf.UnitTests
         public void CacheManager_GetSetFromCache()
         {
             var cacheManager = new CacheManager("Test.Caching.Wpf");
-            // On vide le cache
+            
             cacheManager.Clear();
-
-            // On rajoute les éléments dans le cache :
+            
             for (int i = 1; i < 10; i++)
                 cacheManager.Set("key_" + i.ToString(), i * i, durationCache, "keys");
-
-            Thread.Sleep(5000);
-
-            // On va vérifier maintenant que les éléments y sont bien, et que la masterkey existe bien :
+            
+            Thread.Sleep(2000);
+            
             Assert.IsTrue(cacheManager.IsSet("keys"));
-
+            
             for (int i = 1; i < 10; i++)
             {
                 Assert.IsTrue(cacheManager.IsSet("key_" + i.ToString()));
@@ -38,18 +36,17 @@ namespace MvvX.Plugins.Caching.Wpf.UnitTests
         public void CacheManager_GetWithSet()
         {
             var cacheManager = new CacheManager("Test.Caching.Wpf");
-            // On vide le cache
+            
             cacheManager.Clear();
-
-            // On rajoute les éléments dans le cache :
+            
             for (int i = 1; i < 10; i++)
                 cacheManager.Get("GetWithSetkey_" + i.ToString(), "GetWithSetkeys", () =>
                 {
                     return i * i;
                 });
-
-            // On va vérifier maintenant que les éléments y sont bien, et que la masterkey existe bien :
+            
             Assert.IsTrue(cacheManager.IsSet("GetWithSetkeys"));
+            
             for (int i = 1; i < 10; i++)
             {
                 Assert.IsTrue(cacheManager.IsSet("GetWithSetkey_" + i.ToString()));
@@ -61,10 +58,9 @@ namespace MvvX.Plugins.Caching.Wpf.UnitTests
         public void CacheManager_GetWithSetCacheTime()
         {
             var cacheManager = new CacheManager("Test.Caching.Wpf");
-            // On vide le cache
+            
             cacheManager.Clear();
-
-            // On rajoute les éléments dans le cache :
+            
             for (int i = 1; i < 10; i++)
             {
                 int cacheValue = cacheManager.Get("GetWithSetCacheTimekey_" + i.ToString(), 25, "GetWithSetCacheTimekeys", () =>
@@ -84,17 +80,14 @@ namespace MvvX.Plugins.Caching.Wpf.UnitTests
         public void CacheManager_ClearCache()
         {
             var cacheManager = new CacheManager("Test.Caching.Wpf");
-            // On vide le cache
+            
             cacheManager.Clear();
-
-            // On rajoute les éléments dans le cache :
+            
             for (int i = 1; i < 10; i++)
                 cacheManager.Set("key_" + i.ToString(), i * i, durationCache, "keys");
-
-            // On vide le cache
+            
             cacheManager.Clear();
-
-            // On vérifie que les clés n'y sont plus
+            
             Assert.IsFalse(cacheManager.IsSet("keys"));
             for (int i = 1; i < 10; i++)
                 Assert.IsFalse(cacheManager.IsSet("key_" + i.ToString()));
@@ -104,19 +97,17 @@ namespace MvvX.Plugins.Caching.Wpf.UnitTests
         public void CacheManager_TestPatternKey()
         {
             var cacheManager = new CacheManager("Test.Caching.Wpf");
-            // On vide le cache
-            cacheManager.Clear();
 
-            // On rajoute les éléments dans le cache :
+            cacheManager.Clear();
+            
             for (int i = 1; i < 10; i++)
                 cacheManager.Set("key_" + i.ToString(), i * i, durationCache, "keys");
 
             cacheManager.Remove("keys");
 
-            // On vérifie que les clés n'y sont plus
-            Assert.IsFalse(cacheManager.IsSet("keys"));
+            Assert.IsFalse(cacheManager.IsSet("keys"), "keys is set, not removed");
             for (int i = 1; i < 10; i++)
-                Assert.IsFalse(cacheManager.IsSet("key_" + i.ToString()));
+                Assert.IsFalse(cacheManager.IsSet("key_" + i.ToString()), "key_" + i.ToString() + " is set, not removed");
         }
     }
 }
