+++ MvvX.Plugins.Caching.Wpf.UnitTests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MvvmCross" version="6.2.3" targetFramework="net461" />
+  <package id="System.Console" version="4.3.1" targetFramework="net461" />
+</packages>
