+++ MvvX.Plugins.Caching/Interfaces/ICacheManager.cs
