--- MvvX.Plugins.Caching/PluginLoader.cs
@@ -1,16 +1,0 @@
-﻿using MvvmCross.Platform;
-using MvvmCross.Platform.Plugins;
-
-namespace MvvX.Plugins.Caching
-{
-    public class PluginLoader : IMvxPluginLoader
-    {
-        public static readonly PluginLoader Instance = new PluginLoader();
-
-        public void EnsureLoaded()
-        {
-            var manager = Mvx.Resolve<IMvxPluginManager>();
-            manager.EnsurePlatformAdaptionLoaded<PluginLoader>();
-        }
-    }
-}
