+++ MvvX.Plugins.Caching/PreserveAttribute.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace MvvX.Plugins.Caching
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    class PreserveAttribute : Attribute
+    {
+        public PreserveAttribute() { }
+        public bool AllMembers { get; set; }
+        public bool Conditional { get; set; }
+    }
+}
