+++ MvvX.Plugins.Caching/platforms/android/System.Runtime.Caching/CacheItem.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+
+namespace MvvX.Plugins.Caching.System.Runtime.Caching
+{
+    internal class CacheItem
+    {
+        public string CacheKey { get; internal set; }
+
+        public object Value { get; internal set; }
+
+        public CacheItem(string cacheKey, object value)
+        {
+            CacheKey = cacheKey;
+            Value = value;
+        }
+    }
+}
