+++ MvvX.Plugins.Caching/platforms/ios/System.Runtime.Caching/CacheEntryChangeMonitor.cs
@@ -1,0 +1,10 @@
+
+using System.Collections.Generic;
+
+namespace MvvX.Plugins.Caching.System.Runtime.Caching
+{
+    internal class CacheEntryChangeMonitor
+    {
+        public IList<string> DependencyKeys { get; set; }
+    }
+}
