+++ MvvX.Plugins.Caching/platforms/wpf/PluginConfiguration.cs
@@ -1,0 +1,12 @@
+﻿using MvvmCross.Plugin;
+
+namespace MvvX.Plugins.Caching
+{
+    public class PluginConfiguration : IMvxPluginConfiguration
+    {
+        /// <summary>
+        /// Define the key use to be the global cache key
+        /// </summary>
+        public string GlobalCacheKey { get; set; }
+    }
+}
