+++ appveyor.yml
@@ -1,24 +1,36 @@
-version: 1.6.{build}
+version: '{build}'
+image: Visual Studio 2017
 branches:
   only:
   - master
+skip_tags: true
 max_jobs: 1
 configuration: Release
+build_script:
+- ps: .\build.ps1
+deploy:
+- provider: NuGet
+  disable_publish_on_pr: true
+  api_key:
+    secure: bsq7k6KBk+iI9LutJ6wQKuJr3P5qf0sEfgD05BGK31V6DXQYfLWrEk9Dpx2zw+7a
+  on:
+    branch: master
 assembly_info:
   patch: true
   file: '**\AssemblyInfo.*'
-  assembly_version: '{version}.0'
-  assembly_file_version: '{version}.0'
-  assembly_informational_version: '{version}.0-alpha'
-environment:
-  NuGetApiKey:
-    secure: bsq7k6KBk+iI9LutJ6wQKuJr3P5qf0sEfgD05BGK31V6DXQYfLWrEk9Dpx2zw+7a
-before_build:
-- cmd: >-
-    nuget sources update -Name nuget.org -Source https://api.nuget.org/v3/index.json
-    
-    nuget restore MvvX.Plugins.Caching.sln -verbosity detailed
-build:
-  verbosity: minimal
-on_success:
-- ps: '& "nuspec\packageGenerator.ps1"'
+  assembly_version: '{version}.{build}'
+  assembly_file_version: '{version}.{build}'
+  assembly_informational_version: '{version}.{build}'
+dotnet_csproj:
+  patch: true
+  file: '**\*.csproj'
+  version: '{version}.{build}'
+  package_version: '{version}.{build}'
+  assembly_version: '{version}.{build}'
+  file_version: '{version}.{build}'
+  informational_version: '{version}.{build}'
+install:
+- ps: >-
+    $version = Get-Date -Format "yyyy.1MMdd.1HHmm"
+
+    Update-AppveyorBuild -Version "$version"
