+++ build.cake
@@ -1,0 +1,128 @@
+#tool "nuget:?package=GitVersion.CommandLine"
+#tool "nuget:?package=vswhere"
+
+#addin "nuget:?package=Cake.Figlet"
+
+var sln = new FilePath("./MvvX.Plugins.Caching.sln");
+var outputDir = new DirectoryPath("./artifacts");
+var target = Argument("target", "Default");
+var configuration = Argument("configuration", "Release");
+var verbosityArg = Argument("verbosity", "Minimal");
+var verbosity = Verbosity.Minimal;
+
+var isRunningOnAppVeyor = AppVeyor.IsRunningOnAppVeyor;
+var isPullRequest = AppVeyor.Environment.PullRequest.IsPullRequest;
+
+var version = AppVeyor.Environment.Build.Version;
+Setup(context => {
+
+    var cakeVersion = typeof(ICakeContext).Assembly.GetName().Version.ToString();
+
+    Information(Figlet("Caching"));
+    Information("Building version {0}, ({1}, {2}) using version {3} of Cake.",
+        version,
+        configuration,
+        target,
+        cakeVersion);
+
+    verbosity = (Verbosity) Enum.Parse(typeof(Verbosity), verbosityArg, true);
+});
+
+Task("Clean")
+	.Does(() =>
+{
+    CleanDirectories("./**/bin");
+    CleanDirectories("./**/obj");
+	CleanDirectories(outputDir.FullPath);
+
+	EnsureDirectoryExists(outputDir);
+});
+
+FilePath msBuildPath;
+Task("ResolveBuildTools")
+    .WithCriteria(() => IsRunningOnWindows())
+    .Does(() => 
+{
+    var vsWhereSettings = new VSWhereLatestSettings
+    {
+        IncludePrerelease = true,
+        Requires = "Component.Xamarin"
+    };
+    
+    var vsLatest = VSWhereLatest(vsWhereSettings);
+    msBuildPath = (vsLatest == null)
+        ? null
+        : vsLatest.CombineWithFilePath("./MSBuild/15.0/Bin/MSBuild.exe");
+
+    if (msBuildPath != null)
+        Information("Found MSBuild at {0}", msBuildPath.ToString());
+});
+
+Task("Restore")
+    .IsDependentOn("ResolveBuildTools")
+    .Does(() => 
+{
+    var settings = GetDefaultBuildSettings()
+        .WithTarget("Restore");
+    MSBuild(sln, settings);
+});
+
+Task("Build")
+    .IsDependentOn("ResolveBuildTools")
+    .IsDependentOn("Clean")
+    .IsDependentOn("Restore")
+    .Does(() => 
+{
+
+    var settings = GetDefaultBuildSettings()
+        .WithProperty("Version", version)
+        .WithProperty("PackageVersion", version)
+        .WithProperty("InformationalVersion", version)
+        .WithProperty("NoPackageAnalysis", "True")
+        .WithTarget("Build");
+	
+    MSBuild(sln, settings);
+});
+
+Task("UploadAppVeyorArtifact")
+	.IsDependentOn("CopyPackages")
+	.WithCriteria(() => !isPullRequest)
+	.WithCriteria(() => isRunningOnAppVeyor)
+	.Does(() => {
+
+	Information("Artifacts Dir: {0}", outputDir.FullPath);
+
+	foreach(var file in GetFiles(outputDir.FullPath + "/*")) {
+		Information("Uploading {0}", file.FullPath);
+		AppVeyor.UploadArtifact(file.FullPath);
+	}
+});
+
+
+Task("CopyPackages")
+    .IsDependentOn("Build")
+    .Does(() => 
+{
+    var nugetFiles = GetFiles("./**/bin/" + configuration + "/**/*.nupkg");
+    CopyFiles(nugetFiles, outputDir);
+});
+
+Task("Default")
+	.IsDependentOn("UploadAppVeyorArtifact")
+	.Does(() => {});
+
+RunTarget(target);
+
+MSBuildSettings GetDefaultBuildSettings()
+{
+    var settings = new MSBuildSettings 
+    {
+        Configuration = configuration,
+        ToolPath = msBuildPath,
+        Verbosity = verbosity,
+        ArgumentCustomization = args => args.Append("/m"),
+        ToolVersion = MSBuildToolVersion.VS2017
+    };
+
+    return settings;
+}
