--- nuspec/BootstrapContent/WpfCachingPluginBootstrap.cs.pp
@@ -1,9 +1,0 @@
-using MvvmCross.Platform.Plugins;
-
- namespace $rootnamespace$.Bootstrap
- {
-    public class CachingPluginBootstrap
-        : MvxPluginBootstrapAction<MvvX.Plugins.Caching.PluginLoader>
-    {
-    }
- }
