--- nuspec/MvvX.Plugins.Caching.nuspec
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
- <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-    <metadata>
-        <id>MvvX.Plugins.Caching</id>
-        <version>{BuildNumberVersion}</version>
-        <title>Caching (MvvmCross plugin)</title>
-        <authors>MACK Mathieu</authors>
-        <owners>MACK Mathieu</owners>
-        <licenseUrl>https://github.com/mathieumack/MvvX.CouchBaseLite</licenseUrl>
-        <projectUrl>https://github.com/mathieumack/MvvX.CouchBaseLite</projectUrl>
-        <requireLicenseAcceptance>false</requireLicenseAcceptance>
-        <description>This package contains the Memory Caching plugin for MvvmCross.</description>
-        <tags>mvvm mvvmcross wpf cache</tags>
-        <dependencies>
-          <group>
-            <!-- Always include regardless of target framework -->
-            <dependency id="MvvmCross.Platform" version="4.2.2" />
-        </group>
-        <group targetFramework="portable-win+sl50+wp8">
-            <!-- Otherwise include the Portable CSL -->
-        </group>
-        </dependencies>  
-    </metadata>
-    <files>     
-        <!-- ALL -->
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.XML" target="lib\portable-net45+sl5+wp8+wpa81+win8+monoandroid+monotouch" />
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.dll" target="lib\portable-net45+sl5+wp8+wpa81+win8+monoandroid+monotouch" />
-                
-        <!-- Windows (WPF) -->
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.XML" target="lib\net45" />
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.dll" target="lib\net45" />
-        <file src="..\MvvX.Plugins.Caching.Wpf\bin\Release\MvvX.Plugins.Caching.Wpf.dll" target="lib\net45" />
-		<file src="BootstrapContent\WpfCachingPluginBootstrap.cs.pp" target="content\net45\Bootstrap\WpfCachingPluginBootstrap.cs.pp" />
-		
-        <!-- Android
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.XML" target="lib\MonoAndroid10" />
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.dll" target="lib\MonoAndroid10" />
-        <file src="..\MvvX.Plugins.Caching.Droid\bin\Release\MvvX.Plugins.Caching.Droid.dll" target="lib\MonoAndroid10" />
-		<file src="..\BootstrapContent\CouchBaseLitePluginBootstrap.cs.pp" target="content\MonoAndroid10\Bootstrap\CouchBaseLitePluginBootstrap.cs.pp" />
-		
-        iOS
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.XML" target="lib\MonoTouch10" />
-        <file src="..\MvvX.Plugins.Caching\bin\Release\MvvX.Plugins.Caching.dll" target="lib\MonoTouch10" />
-        <file src="..\MvvX.Plugins.Caching.Touch\bin\Release\MvvX.Plugins.Caching.Touch.dll" target="lib\MonoTouch10" />
-		<file src="..\BootstrapContent\CouchBaseLitePluginBootstrap.cs.pp" target="content\MonoTouch10\Bootstrap\CouchBaseLitePluginBootstrap.cs.pp" /> -->
-    </files>
-</package>
-
